+++ crates/gqlitedb/src/functions/value.rs
@@ -3,36 +3,6 @@ use crate::{error::RunTimeError, graph};
 use super::{ExpressionType, FResult, FunctionTypeTrait};
 
 #[derive(Debug, Default)]
-pub(super) struct Coalesce {}
-
-impl super::FunctionTrait for Coalesce
-{
-  fn call(&self, arguments: Vec<graph::Value>) -> crate::Result<graph::Value>
-  {
-    for arg in arguments
-    {
-      match arg
-      {
-        graph::Value::Invalid =>
-        {}
-        other => return Ok(other),
-      }
-    }
-    Ok(graph::Value::Invalid)
-  }
-  fn validate_arguments(&self, _: Vec<ExpressionType>) -> crate::Result<ExpressionType>
-  {
-    Ok(ExpressionType::Variant)
-  }
-  fn is_deterministic(&self) -> bool
-  {
-    true
-  }
-}
-
-super::declare_function!(coalesce, Coalesce, custom_trait);
-
-#[derive(Debug, Default)]
 pub(super) struct HasLabel {}
 
 impl HasLabel